Mount instead of build when prev child DNE
Mounting should happen when a component is being rendered for the firs time.
When building DOM components, if a new child is inserted, it should run the
mounting life-cycle rather than the building. (React has this behavior)

TODO, this also means that -build should probably be renamed to -build! because
it can call -mount! and side-effects are allowed there.